[SHM][AUX SHM] Updated tests, print, and static; merge complete - Combined shm and aux shm tests pass - Combined printing process works and prints all values to console - Static processes work (not combined in order for convenience of testing) - Removed shm_wrapper_aux folder
