+++ logger/logger.config
@@ -10,7 +10,7 @@
 LOG_TO_STDOUT: YES
 
 // Output logs to log file? (Yes/No)
-LOG_TO_FILE: YES
+LOG_TO_FILE: NO
 
 // Output logs over the network to Dawn? (Yes/No)
 LOG_TO_NETWORK: YES
