+++ shm_wrapper/print_aux_shm.c
